feat(sale): Parcial version

+++ src/Ambev.DeveloperEvaluation.Application/Sales/Commands/CreateSaleCommand.cs
@@ -10,7 +10,7 @@ public class CreateSaleCommand : IRequest<SaleResult>
     public string SaleNumber { get; set; }
     public DateTime SaleDate { get; set; }
     public decimal TotalAmount { get; set; }
-    public decimal TotalItems { get; set; }
+    public int TotalItems { get; set; }
     public decimal TotalDiscounts { get; set; }
     public string Branch { get; set; }
     public bool IsCancelled { get; set; }
